refactor: 이해성 향상을 위한 패키지 분리
header, body 패키지 생성하고 배분
