+++ src/main/java/webserver/http/response/header/MimeType.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package webserver.http.response.header;
 
 public enum MimeType {
