+++ src/main/java/webserver/http/response/header/ResponseHeader.java
@@ -1,4 +1,4 @@
-package webserver.http.response;
+package webserver.http.response.header;
 
 
 public enum ResponseHeader {
