+++ src/main/java/webserver/http/response/header/StatusLine.java
@@ -1,4 +1,4 @@
-package webserver.http.response;
+package webserver.http.response.header;
 
 import static webserver.http.response.ResponseMessageHeader.BLANK;
