+++ src/test/java/webserver/http/HttpContentTypeTest.java
@@ -3,7 +3,7 @@ package webserver.http;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.http.response.HttpContentType;
+import webserver.http.response.header.HttpContentType;
 
 import java.lang.reflect.Field;
 import java.util.Map;
