+++ src/test/java/webserver/http/ResponseMessageHeaderTest.java
@@ -3,6 +3,7 @@ package webserver.http;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import webserver.http.response.ResponseMessageHeader;
+import webserver.http.response.header.MimeType;
 
 import static org.junit.jupiter.api.Assertions.*;
